Move all generated .ui files to a subfolder
