+++ pyuic.json
@@ -2,7 +2,7 @@
     "files": [
         [
             "randovania/gui/*.ui",
-            "randovania/gui"
+            "randovania/gui/generated"
         ]
     ],
     "hooks": [],
