+++ randovania/gui/.gitignore
@@ -1,2 +1,1 @@
 *.ui.autosave
-*_ui.py
