+++ randovania/gui/game_patches_window.py
@@ -1,6 +1,6 @@
 from PySide2.QtWidgets import QMainWindow, QComboBox
 
-from randovania.gui.game_patches_window_ui import Ui_GamePatchesWindow
+from randovania.gui.generated.game_patches_window_ui import Ui_GamePatchesWindow
 from randovania.interface_common.options import Options
 from randovania.layout.patcher_configuration import PickupModelStyle, PickupModelDataSource
