+++ randovania/gui/item_configuration_popup.py
@@ -6,7 +6,7 @@ from randovania.game_description import default_database
 from randovania.game_description.item.major_item import MajorItem
 from randovania.game_description.resources.resource_type import ResourceType
 from randovania.gui.common_qt_lib import set_default_window_icon
-from randovania.gui.item_configuration_popup_ui import Ui_ItemConfigurationPopup
+from randovania.gui.generated.item_configuration_popup_ui import Ui_ItemConfigurationPopup
 from randovania.layout.major_item_state import MajorItemState
 
 _INVALID_MODELS = {
