+++ randovania/gui/permalink_dialog.py
@@ -1,7 +1,7 @@
 from PySide2.QtWidgets import QDialog, QApplication
 
 from randovania.gui import common_qt_lib
-from randovania.gui.permalink_dialog_ui import Ui_PermalinkDialog
+from randovania.gui.generated.permalink_dialog_ui import Ui_PermalinkDialog
 from randovania.layout.permalink import Permalink
